requirements.txt
streamlit
tensorflow-cpu
numpy
pandas
plotly
pillow
gdown